Removed the sleep changes from this PR. They caused compile issues. I need to think through them more carefully.
