+++ include/boost/thread/thread_only.hpp
@@ -23,5 +23,7 @@
 #if defined BOOST_THREAD_PROVIDES_INTERRUPTIONS
 #include <boost/thread/detail/thread_interruption.hpp>
 #endif
+#include <boost/thread/v2/thread.hpp>
+
 
 #endif
