+++ include/boost/thread/win32/thread_data.hpp
@@ -189,7 +189,6 @@ namespace boost
         {
           interruptible_wait(detail::win32::invalid_handle_value, detail::internal_platform_clock::now() + detail::platform_duration(rel_time));
         }
-
         inline BOOST_SYMBOL_VISIBLE void sleep(system_time const& abs_time)
         {
           const detail::real_platform_timepoint ts(abs_time);
@@ -199,33 +198,6 @@ namespace boost
             d = (std::min)(d, detail::platform_milliseconds(100));
             interruptible_wait(detail::win32::invalid_handle_value, detail::internal_platform_clock::now() + d);
             d = ts - detail::real_platform_clock::now();
-          }
-        }
-#endif
-
-#ifdef BOOST_THREAD_USES_CHRONO
-        template <class Rep, class Period>
-        void sleep_for(const chrono::duration<Rep, Period>& d)
-        {
-          interruptible_wait(detail::win32::invalid_handle_value, detail::internal_platform_clock::now() + detail::platform_duration(d));
-        }
-
-        template <class Duration>
-        void sleep_until(const chrono::time_point<chrono::steady_clock, Duration>& t)
-        {
-          sleep_for(t - chrono::steady_clock::now());
-        }
-
-        template <class Clock, class Duration>
-        void sleep_until(const chrono::time_point<Clock, Duration>& t)
-        {
-          typedef typename common_type<Duration, typename Clock::duration>::type CD;
-          CD d = t - Clock::now();
-          while (d > CD::zero())
-          {
-            d = (std::min)(d, CD(chrono::milliseconds(100)));
-            sleep_for(d);
-            d = t - Clock::now();
           }
         }
 #endif
@@ -240,7 +212,6 @@ namespace boost
           {
             non_interruptible_wait(detail::win32::invalid_handle_value, detail::internal_platform_clock::now() + detail::platform_duration(rel_time));
           }
-
           inline BOOST_SYMBOL_VISIBLE void sleep(system_time const& abs_time)
           {
             const detail::real_platform_timepoint ts(abs_time);
@@ -250,33 +221,6 @@ namespace boost
               d = (std::min)(d, detail::platform_milliseconds(100));
               non_interruptible_wait(detail::win32::invalid_handle_value, detail::internal_platform_clock::now() + d);
               d = ts - detail::real_platform_clock::now();
-            }
-          }
-#endif
-
-#ifdef BOOST_THREAD_USES_CHRONO
-          template <class Rep, class Period>
-          void sleep_for(const chrono::duration<Rep, Period>& d)
-          {
-            non_interruptible_wait(detail::win32::invalid_handle_value, detail::internal_platform_clock::now() + detail::platform_duration(d));
-          }
-
-          template <class Duration>
-          void sleep_until(const chrono::time_point<chrono::steady_clock, Duration>& t)
-          {
-            sleep_for(t - chrono::steady_clock::now());
-          }
-
-          template <class Clock, class Duration>
-          void sleep_until(const chrono::time_point<Clock, Duration>& t)
-          {
-            typedef typename common_type<Duration, typename Clock::duration>::type CD;
-            CD d = t - Clock::now();
-            while (d > CD::zero())
-            {
-              d = (std::min)(d, CD(chrono::milliseconds(100)));
-              sleep_for(d);
-              d = t - Clock::now();
             }
           }
 #endif
